Dont use silly ids, actually create the data before test runs
